+++ record/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.2.0
+* feat: Add recorder state streams.
+* feat: Add recorder amplitude stream.
+* core: Update dependencies.
+
 ## 4.1.1
 * core: README.md updates.